requirements.txt
matplotlib
numpy
Pillow
torch
pandas
torchvision
tensorboard
pyyaml
